break iterator and iterable out from query
