+++ src/main/java/org/sql2o/tools/ResultSetUtils.java
@@ -1,0 +1,20 @@
+package org.sql2o.tools;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+/**
+ * Utilities for {@link java.sql.ResultSet}.
+ */
+public class ResultSetUtils {
+    public static Object getRSVal(ResultSet rs, int idx) throws SQLException {
+        Object o = rs.getObject(idx);
+        // oracle timestamps are not always convertible to a java Date. If ResultSet.getTimestamp is used instead of
+        // ResultSet.getObject, a normal java.sql.Timestamp instance is returnd.
+        if (o != null && o.getClass().getCanonicalName().startsWith("oracle.sql.TIMESTAMP")){
+            o = rs.getTimestamp(idx);
+        }
+
+        return o;
+    }
+}
